refactor: . Move to src/Obsidian/ most top-level files that integrate with Obsidian
QueryRenderer.ts not yet moved, as I'm unsure where the other
Layout and/or Renderer will go.
