+++ src/Api/createTaskLineModalHelper.ts
@@ -1,7 +1,7 @@
 import type { App } from 'obsidian';
 import type { Task } from '../Task/Task';
 import { taskFromLine } from '../Commands/CreateOrEditTaskParser';
-import { TaskModal } from '../TaskModal';
+import { TaskModal } from '../Obsidian/TaskModal';
 import type { ITaskModal, taskModalFactory } from './createTaskLineModal';
 
 /**
