+++ src/Commands/CreateOrEdit.ts
@@ -1,5 +1,5 @@
 import { App, Editor, MarkdownView, View } from 'obsidian';
-import { TaskModal } from '../TaskModal';
+import { TaskModal } from '../Obsidian/TaskModal';
 import type { Task } from '../Task/Task';
 import { DateFallback } from '../DateFallback';
 import { taskFromLine } from './CreateOrEditTaskParser';
