+++ src/Obsidian/InlineRenderer.ts
@@ -1,11 +1,11 @@
 import type { MarkdownPostProcessorContext, Plugin } from 'obsidian';
 import { MarkdownRenderChild } from 'obsidian';
-import { GlobalFilter } from './Config/GlobalFilter';
-import { TaskLayoutOptions } from './Layout/TaskLayoutOptions';
-import { QueryLayoutOptions } from './QueryLayoutOptions';
-import { Task } from './Task/Task';
-import { TaskLineRenderer } from './TaskLineRenderer';
-import { TaskLocation } from './Task/TaskLocation';
+import { GlobalFilter } from '../Config/GlobalFilter';
+import { TaskLayoutOptions } from '../Layout/TaskLayoutOptions';
+import { QueryLayoutOptions } from '../QueryLayoutOptions';
+import { Task } from '../Task/Task';
+import { TaskLineRenderer } from '../TaskLineRenderer';
+import { TaskLocation } from '../Task/TaskLocation';
 
 export class InlineRenderer {
     constructor({ plugin }: { plugin: Plugin }) {
