+++ src/Obsidian/TasksEvents.ts
@@ -1,8 +1,8 @@
 import type { EventRef, Events as ObsidianEvents } from 'obsidian';
 
+import type { Task } from '../Task/Task';
+import { logging } from '../lib/logging';
 import type { State } from './Cache';
-import type { Task } from './Task/Task';
-import { logging } from './lib/logging';
 
 enum Event {
     CacheUpdate = 'obsidian-tasks-plugin:cache-update',
