+++ src/ui/Menus/TaskEditingMenu.ts
@@ -1,6 +1,6 @@
 import { Menu, type MenuItem } from 'obsidian';
 import type { Task } from '../../Task/Task';
-import { replaceTaskWithTasks } from '../../File';
+import { replaceTaskWithTasks } from '../../Obsidian/File';
 import type { TaskEditingInstruction } from '../EditInstructions/TaskEditingInstruction';
 
 /**
